Add tests for naming when not using the plugin
