+++ no-plugin-tests/src/test/scala/chiselTests/BetterNamingTests.scala
@@ -1,0 +1,1 @@
+../../../../../src/test/scala/chiselTests/BetterNamingTests.scala
