+++ no-plugin-tests/src/test/scala/chiselTests/CatSpec.scala
@@ -1,0 +1,1 @@
+../../../../../src/test/scala/chiselTests/util/CatSpec.scala
