+++ no-plugin-tests/src/test/scala/chiselTests/CustomBundle.scala
@@ -1,0 +1,1 @@
+../../../../../src/test/scala/chiselTests/CustomBundle.scala
