+++ no-plugin-tests/src/test/scala/chiselTests/NamingAnnotationTest.scala
@@ -1,0 +1,1 @@
+../../../../../src/test/scala/chiselTests/NamingAnnotationTest.scala
